Do not perform SSH stuff in .bashrc
SSH / keys are a confusing item in GNU/Linux. GNOME manages them in
its own way. Managing them from shells assumes everything is launched
from the shell, which in my case it is not. You can also start agents
using systemd, but I guess that messes with GNOME and the confusion
circles round.

For now, it seems to WFM in GNOME, so I'll Occam this shell file.