Automatically add SSH keys
This means they will be cached for the uptime of a machine.

I added the following, all copied from /etc/xdg/autostart to turn off
the GNOME equivalents. I'd rather use the vanilla ssh-agent (which is
started by XFCE automatically)

  /home/cht/.config/autostart:
  total used in directory 24 available 294017316
  drwxr-xr-x  2 cht cht 4096 Jan 26 13:40 .
  drwx------ 68 cht cht 4096 Jan 25 11:04 ..
  -rw-r--r--  1 cht cht  520 Jan 26 13:35 gnome-keyring-pkcs11.desktop
  -rw-r--r--  1 cht cht  512 Jan 26 13:35 gnome-keyring-secrets.desktop
  -rw-r--r--  1 cht cht  479 Jan 26 13:35 gnome-keyring-ssh.desktop

The change made to each of these files is

X-GNOME-Autostart-enabled=false

To, well, stop them!